Move MemberRow to separate file
